build: Generate directory and ZIP archive alongside sdcard.img
When updating an existing mountable SD card it may be useful to have
access to the SD card contents as a local folder, since unpacking
individual files from an SD card image is mildly tedious.

Assemble the SD card contents in a local folder and use this to
construct both the raw disk image file sdcard.img and a ZIP archive
sdcard.zip.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>